Python_bootcamp/day06/src/requirements.txt
grpcio
alembic
pydantic
sqlalchemy
psycopg2
grpcio-tools
googleapis-common-protos